% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Construct coded name}
\usage{
code(...)
}
\arguments{
\item{...}{character string components of a name.}
}
\value{
The coded variable name.
}
\description{
Construct coded name
}
\examples{
code("GF", "A20to24")

}

