% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{enter_api_key}
\alias{enter_api_key}
\title{Prompt user to update ASC API key.}
\usage{
enter_api_key(file = "key.Rda")
}
\arguments{
\item{file}{The filename for the installed ACS API key}
}
\description{
Prompt user to update ASC API key.
}
\seealso{
\code{\link{load_api_key}} for loading from file.
}

