% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{have_api_key}
\alias{have_api_key}
\title{Test if ACS API is installed.}
\usage{
have_api_key(file = "key.Rda")
}
\arguments{
\item{file}{The filename for the installed ACS API key}
}
\description{
Test if ACS API is installed.
}
\seealso{
\code{\link{load_api_key}} for loading from file,
  and \code{\link{enter_api_key}} for entering with interactive prompt.
}

