% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{load_api_key}
\alias{load_api_key}
\title{Load ASC API key from file.}
\usage{
load_api_key(source_file, file = "key.Rda")
}
\arguments{
\item{source_file}{A file containing an ACS API key}

\item{file}{The filename for the installed ACS API key}
}
\description{
Load ASC API key from file.
}
\seealso{
\code{\link{enter_api_key}} for entering with interactive prompt.
}

