% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{retrieve_api_key}
\alias{retrieve_api_key}
\title{Retrieve installed ACS API key}
\usage{
retrieve_api_key(file = "key.Rda")
}
\arguments{
\item{file}{The filename for the installed ACS API key}
}
\description{
Retrieve installed ACS API key
}

