% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_var.R
\name{table_var}
\alias{table_var}
\title{Retrieve table variable given coded name}
\usage{
table_var(strs, end_year, table)
}
\arguments{
\item{strs}{character vector.}

\item{end_year}{end year of survey.}

\item{table}{table name.}
}
\value{
The full variable name (e.g., "B17006_004").
}
\description{
Retrieve table variable given coded name
}
\examples{
table_var(c("SubSahAfr", "Gha"), end_year = 2010, table = "B04001")

}

